xform: move select.opt functions to select_funcs.go
This is part of a refactoring to split up xform custom functions,
similar to normalization custom functions.

Release note: None
